+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Fixed
 
 * Correctly update theme colours on change again
 * Imprint is now called "Imprint" and not "Impress".
+* Logo files weren't uploaded to public namespace.
 
 `2.3.1`_ – 2021-12-17
 ---------------------